refactor: rename s3.py to s3_shared_context
That makes it more explicit what the intention is as there will also be code for generating specific user context which will be added later on.
